tts cleanup, fixed import in overlay
